requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
Werkzeug==3.0.1
SQLAlchemy==2.0.25
Pillow==10.2.0
python-socketio==5.10.0
requests==2.31.0
feedparser==6.0.10
python-dotenv==1.0.0
bcrypt==4.1.2
APScheduler==3.10.4
eventlet==0.33.3
